Scripting to create the CD and analog replays
The System Explorer VIs are in a somewhat weird state, as some of them
use the Advanced SysDef API while others don't. I believe we will
need to transition to using this everywhere, in order to properly
support running the scripting API through built apps and TestStand, but
further research is required.