Update RIO <-> APU mappings and show in APU list
These are persisted in the SysDef as RIO -> APU, even though APU -> RIO might appear to make more sense based on how the information is displayed. This is because the APU sections can be removed, but we still need to know the mapping, so we can't store the information under the APU sections. Upon further reflection, we could store them outside of the APU sections (e.g. adjacent to their current location), but then we would potentially have to rewrite every property when a bitfile is added/removed, instead of modifying just one.

It's not immediately clear which format is best.