Only read the multi-fpga feature toggle from the ini key in On Load
When opening pages, we should use the multi-fpga key stored in the system definition. This prevents us from getting into weird states if the ini file is changed while the system explorer is open, and should allow us to use the multi-fpga codepaths in tests without setting the ini key.