HACK: Rename colliding files during build.
The GXML library uses deprecated versions of the NI Data Type library. The custom device uses the supported version. During the build, some of the names collide, causing extra folders to be placed on disk. We should figure out a way to update the GXML library to use the supported VIs and then pull this hack.